plplay: decode_loop(): Ignore decoder errors
Ignore EIO from avcodec_send_packet() in decode_loop(). This happens
because the receiver doesn't see the very first frame that the remote
encoder puts out, so the receiver has to wait until it sees an I-frame.
There will be a burst of frames at the beginning that aren't decodeable.