plplay: decode_loop(): Call av_hwframe_map() on decoded frames to test validity
At least on my Intel Arc A310, it's apparently possible for hardware
decoded frames to not be fully decoded (or the results of decoding are
lazily reported?) until an actual underlying call to vaSyncSurface() is
made as part of attempting to display the frame contents.

That means a corrupt, undecodable VAAPI frame may not be detected until
the render loop calls pl_map_avframe_ex(), which then calls an internal
function pl_map_avframe_derived(), which calls av_hwframe_map(), which
calls vaSyncSurface(), which fails. That results in this ffmpeg log
message:

[AVHWFramesContext @ …] Failed to sync surface 0x15: 23 (internal decoding error).

By making a call to av_hwframe_map() in the decoder loop that is
equivalent to what gets called by pl_map_avframe_derived() in the render
loop, the decoder loop can determine if the frame is invalid and discard
it rather than queueing it for the render loop to choke on later.

It would be nice to save the AV_PIX_FMT_DRM_PRIME result of calling the
new function try_avframe_derive() from the decoder loop so that the same
work isn't repeated by pl_map_avframe_derived() in the render loop but I
ran into memory corruption and didn't figure it out yet. I'm not sure
how expensive the VAAPI → DRM_PRIME format conversion is, though.

[Aside: what pl_map_avframe_ex() → pl_map_avframe_derived() is trying to
do is to derive a replacement AVFrame in the AV_PIX_FMT_DRM_PRIME format
from the original AVFrame in the AV_PIX_FMT_VAAPI format.

An AV_PIX_FMT_DRM_PRIME format AVFrame contains "DRM-managed buffers
exposed through PRIME buffer sharing" [0], which "can be imported
directly by EGL using the EGL_EXT_image_dma_buf_import and
EGL_EXT_image_dma_buf_import_modifiers extensions" [1].

According to the kernel docs [2], "PRIME is the cross device buffer
sharing framework in drm, originally created for the OPTIMUS range of
multi-gpu platforms. To userspace PRIME buffers are dma-buf based file
descriptors." And, dma-buf [3] "provides the framework for sharing
buffers for hardware (DMA) access across multiple device drivers and
subsystems, and for synchronizing asynchronous hardware access. As an
example, it is used extensively by the DRM subsystem to exchange buffers
between processes, contexts, library APIs within the same process, and
also to exchange buffers with other subsystems such as V4L2."

So if I understand this all correctly, the goal is to keep the decoded
frame content in GPU memory after it comes out of the VAAPI decoder and
render it onto the screen using OpenGL *without* needing to copy it from
GPU memory to CPU memory and back to GPU memory.

[0]: https://ffmpeg.org/doxygen/trunk/pixfmt_8h.html

[1]: https://ffmpeg.org/doxygen/trunk/structAVDRMFrameDescriptor.html

[2]: https://docs.kernel.org/gpu/drm-mm.html#prime-buffer-sharing

[3]: https://docs.kernel.org/driver-api/dma-buf.html