setupRealtimePriority(): Apply SCHED_FIFO priority to all threads
Not really all, but almost all. This just sets the default for
pthread_create() with a NULL `attr` argument. Unfortunately
pthread_setattr_default_np() is nonportable.

If ffmpeg is built with --enable-pocketsphinx (like on Debian unstable
right now) it will pull in libopenblas which sneaks an entire thread
pool into the process via an __attribute__((constructor)) function. See
https://bugs.debian.org/cgi-bin/bugreport.cgi?bug=1059505. But vacon
also doesn't care about the priority of threads in some indirect
dependency that we're never going to use.