NetworkHandler: Add a class for the WebRTC logic
This adds a NetworkHandler class to set up the WebRTC connection,
talking to the signaling server and setting up the peer connection,
using the libdatachannel library.

This unifies the slightly different logic for each side of the SDP
offer/answer that was previously split across the old camera_sender
(offer sender, answer receiver) and video_receiver (offer receiver,
answer sender).

Untested. This sets up a video track using
rtc::Description::Direction::SendRecv rather than ::SendOnly on one side
and ::RecvOnly on the other, and I'm not sure if it's correct to chain
the rtc::RtcpReceivingSession to the packetizer (but it compiles).

Otherwise two peers running this NetworkHandler and connecting to the
new signaling server seem to be successfully bringing up the WebRTC
connection (SDP, SRTP, etc.) between them, but without any video flowing
yet.