RtpDepacketizer: Add a class to wrap ffmpeg's SDP/RTP decoder
This adds an RtpDepacketizer class that wraps ffmpeg's SDP/RTP decoder,
using custom AVIOContext* objects to provide the SDP and RTP input. It
is currently hardcoded with a fake SDP descriptor to handle an H.265
media stream on dynamic payload type 96, this should be generalized if
we continue using ffmpeg's RTP decoder implementation (we should not).

Incoming RTP packets are provided externally by calling
submitRtpPacket().

There is currently no way to read the output of ffmpeg's RTP decoder.
The easy thing to do would be to expose the RtpDepacketizer's
AVFormatContext* member.