linux::Encoder::InitMfxVideoParamVpp(): Set mfxFrameInfo.Shift = 1
P010 is a planar 4:2:0 format that stores 10 bits per channel. Each 10
bit sample is stored in a 2 byte word, with 6 bits necessarily
wasted.

Apparently the Intel® Video Processing Library also requires the
mfxFrameInfo::Shift field to be set to 1 if the FourCC is P010. This is
in addition to the completely redundant BitDepthChroma and BitDepthLuma
fields whose values are already fully implied by the FourCC (the "10" in
"P010").

This change seems to substantially change the performance of the
VPP/ENCODE pipeline. Prior to this commit, I was getting ~8200 us per
frame for scaling + encoding (range ~7800 to ~23000 us) of 1920x1080
content. With this commit I'm getting ~5900 us average (range ~5600 to
~18000 us).