linux::Encoder: Merge VPP/ENCODE parameter setting into InitMfxVideoParams()
Also add a SetMfxFourCc() helper that sets up the various FourCC related
parameters and handles mapping between the camera input pixel format and
the encoder pixel format (i.e. the 4:2:0 formats get mapped to P010, the
4:2:2 formats get mapped to Y210).

Remove the consistency check in EncodeCameraFrame() because apparently
the standard Linux FourCC "YUYV" (packed 4:2:2 YUV) is alternatively
known as "YUY2" in the Wintel world.