linux::Encoder: Import raw camera frame data by reference
I could not find anything in the latest version of the libvpl API
reference about the semantics of mfxFrameSurface1 and whether they can
be constructed by the caller or whether they can only be obtained
through the various MFXMemory_GetSurfaceFor…() functions.

However, according to the "Intel Media Developer's Guide" [0] from 2016,
which covers the old Intel® Media SDK library, on page 88:

    Surface buffer layout

    System memory allocations also need to explicitly match data layout
    expectations for the format used. While the examples below are for input
    surfaces, output surfaces can also be set up the same way.

    NV12

    bitsPerPixel = 12; // NV12, YV12, and I420 formats are 12 bits per pixel
    surfaceSize = width * height * bitsPerPixel / 8;
    surfaceBuffersIn = (mfxU8*) new mfxU8[surfaceSize * nVPPSurfNumIn];
    pVPPSurfacesIn = new mfxFrameSurface1 *[nVPPSurfNumIn];
    MSDK_CHECK_POINTER(pVPPSurfacesIn, MFX_ERR_MEMORY_ALLOC);
    for (int i = 0; i < nVPPSurfNumIn; i++) {
        pVPPSurfacesIn[i] = new mfxFrameSurface1;
        memset(pVPPSurfacesIn[i], 0, sizeof(mfxFrameSurface1));
        memcpy(&(pVPPSurfacesIn[i]->Info), &(VPPParams.vpp.In), sizeof(mfxFrameInfo));
        pVPPSurfacesIn[i]->Data.Y = &surfaceBuffersIn[surfaceSize * i];
        pVPPSurfacesIn[i]->Data.UV = pVPPSurfacesIn[i]->Data.Y + width * height;

        pVPPSurfacesIn[i]->Data.Pitch = width;
    }

This convoluted code sample seems to be saying that you can indeed fake
an mfxFrameSurface1 by zero-initializing it and copying in the relevant
FrameInfo member from your mfxVideoParam and fill in the FrameData
member's channel pointers and Pitch field.

And this seems to work, at least for MFX_IOPATTERN_IN_SYSTEM_MEMORY.

[0] https://www.intel.com/content/dam/develop/external/us/en/documents/intel-media-developers-guide-720530.pdf