linux::CameraFrame: Re-design to support multiple consumers of camera frame data
Re-design the CameraFrame class so that it can support multiple
consumers of the same camera frame data such as a video encoder as well
as a local self-view.

CameraFrame's are now wrapped in std::shared_ptr and the destructor
VIDIOC_QBUF's the buffer back to the kernel. So the existence of a
CameraFrame keeps the buffer in the userspace application.

The CameraFrame gets a copy of the 'struct v4l2_buffer' that the frame
was VIDIOC_DQBUF'd from and a pointer to the camera frame data, but the
mmap() / munmap() for the frame buffer now needs to be managed
externally.

A pointer to the Camera that captured the frame is needed so the
CameraFrame can tell consumers the 'struct v4l2_pix_format'
corresponding to the frame which has critical information like the
width, height, pitch, and actual pixel format selected by the V4L2
driver (which may differ from the requested pixel format).

This requires corresponding changes to linux::Camera and linux::Encoder.