Revise linux::Camera
Significantly revise linux::Camera based on the `v4l-dmabuf-test`
experiments. Callers will need to be updated.

Get rid of the CameraFrame class. Introduce CameraBuffer and
CameraBufferRef classes.

CameraBuffer holds all of the resources and metadata associated with an
individual V4L2 camera buffer. Data such as the OpenGL texture and the
mmap()'d frame buffer, metadata such as the DQBUF'd v4l2_buffer struct
with current sequence number, timestamp, etc. CameraBuffer's are created
only at startup and held in the Camera class.

When a camera buffer is DQBUF'd from the kernel, a
std::shared_ptr<CameraBufferRef> is created that holds a reference to
the actual CameraBuffer. The CameraBufferRef can be shared between
several consumers and when destroyed it QBUF's the buffer back to the
kernel.

Camera exports a dmabuf fd from the V4L2 driver as well as mmap()'ing
the camera frame data, since we will need both.

The Camera::ExportBuffersToOpenGL() method uses the
EGL_EXT_image_dma_buf_import and GL_OES_EGL_image_external extensions to
import the V4L2 dmabuf fd into an OpenGL texture. This must be called
with an appropriately prepared SDL_Renderer*. I believe this method
needs to be called from the render thread.