Add classes GenericRtpDepacketizer, GenericRtpPacketizer
This commit adds a simple implementation of a generic RTP packetizer/de-packetizer that is compatible with the libdatachannel MediaHandler API. It is sort of based on the H264/H265 packetizer/de-packetizer implementations in libdatachannel.

Each video packet passed to the packetizer is split into fragments small enough to be carried via RTP messages. There is a simple one-byte header in front of each fragment payload that indicates whether the fragment is from the beginning, middle, or end of the packet.

The de-packetizer reassembles the fragments into the original packet, but it will truncate the packet, leaving only the contiguous head, if any fragments are missing.

This should be good enough for Vacon's use cases to replace the video codec-specific packetizers/de-packetizers.