NetworkHandler: Perform bi-directional video codec negotiation
There are four codecs to be selected:

* Encoder on the offer side.
* Decoder on the offer side.
* Encoder on the answer side.
* Decoder on the answer side.

The encoder on the offer side needs to match the decoder on the answer
side.

The decoder on the offer side needs to match the encoder on the answer
side.

Every supported codec on the offer side will be included in the offer,
as SendRecv (if encoder+decoder support), SendOnly (if encoder support),
or RecvOnly (if decoder support) video tracks. On the answer side the
offerer's capabilities will be matched against the local capabilities
and either a single SendRecv video track or a SendOnly and a RecvOnly
video track will be included in the answer.

The results of the NetworkHandler's video codec negotiation need to be
wired up to encoder/decoder initialization so the selected codecs are
actually used.