linux::Camera: Reorder ChromaFormat to prefer 4:2:0 over 4:2:2
Because we only use codecs that produce 4:2:0 output, there is no real
advantage to capturing in 4:2:2 and scaling it down to 4:2:0 before it
gets to the encoder. So if a 4:2:0 format is available from the camera
(i.e., NV12, which is apparently the only pixel format that the Intel
encoder can accept without requiring VPP scaling), use it in preference
to any 4:2:2 pixel formats. Of course, some cameras like the Elgato
Facecam only have 4:2:2 (UYVY) available and we'll have to use the 4:2:2
format. Other cameras like the Razer Kiyo Pro and Razer Kiyo Pro Ultra
can do both 4:2:0 (NV12) and 4:2:2 (YUYV).